[astups] Track and list several tag
The main now lists the visible markers and give their ID.

The CMakeLists.txt no longer copy the content of astups/data,
to have the executable working, one need to create a symlink.

TODO: Write documentation on how to write and format the content
of the "cfg" file used by the MutliTagTracker.